requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.15.0
pillow>=8.0.0
gradio>=3.8.0
numpy>=1.19.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
huggingface_hub>=0.10.0
requests>=2.25.0 